requirements.txt
# ===================================
# PROJET IA - EXTRACTION & ANALYSE
# ===================================

# --- Extraction de base ---
python-pptx>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0

# --- Extraction avancee avec Unstructured.io ---
unstructured[all-docs]>=0.18.0

# --- OCR (Optical Character Recognition) ---
pytesseract>=0.3.10
pillow>=10.0.0
pdf2image>=1.17.0
opencv-python>=4.8.0
numpy>=1.24.0

# --- PDF avance ---
PyMuPDF>=1.23.0

# --- LLM & API (GRATUIT - Groq) ---
groq>=0.4.0
python-dotenv>=1.0.0

# --- RAG & Agent (LangChain + ChromaDB) ---
langchain>=0.1.0
langchain-groq>=0.0.1
chromadb>=0.4.22
sentence-transformers>=2.2.0

# --- Validation & Utils ---
pydantic>=2.0.0
jinja2>=3.1.0

# --- Utilitaires ---
pathlib>=1.0.1
